Added args checking to AuditoryService

+++ src/main/java/services/implementations/AuditoryService.java
@@ -14,7 +14,9 @@ import java.util.stream.Collectors;
 
 @Component
 public class AuditoryService implements Service<AuditoryModel> {
-    public static final String DAO_ERROR = "Dao error";
+    private static final String DAO_ERROR = "Dao error";
+    private static final String NULL_ERROR = "Null is passed";
+    private static final String ID_ERROR = "Invalid id is passed";
 
     private AuditoryDao auditoryDao;
 
@@ -22,12 +24,18 @@ public class AuditoryService implements Service<AuditoryModel> {
     }
 
     @Autowired
-    public AuditoryService(AuditoryDao auditoryDao) {
+    public AuditoryService(AuditoryDao auditoryDao) throws ServiceException {
+        if (auditoryDao == null) {
+            throw new ServiceException(NULL_ERROR);
+        }
         this.auditoryDao = auditoryDao;
     }
 
     @Override
     public AuditoryModel save(AuditoryModel item) throws ServiceException {
+        if (item == null) {
+            throw new ServiceException(NULL_ERROR);
+        }
         Auditory auditoryEntity = new Auditory();
         auditoryEntity.setLocation(item.getLocation());
         try {
@@ -40,6 +48,9 @@ public class AuditoryService implements Service<AuditoryModel> {
 
     @Override
     public void deleteById(int id) throws ServiceException {
+        if (id <= 0) {
+            throw new ServiceException(ID_ERROR);
+        }
         try {
             auditoryDao.deleteById(id);
         } catch (DaoException e) {
@@ -49,6 +60,9 @@ public class AuditoryService implements Service<AuditoryModel> {
 
     @Override
     public AuditoryModel findById(int id) throws ServiceException {
+        if (id <= 0) {
+            throw new ServiceException(ID_ERROR);
+        }
         try {
             Auditory auditoryEntity = auditoryDao.findById(id);
             return new AuditoryModel(auditoryEntity.getId(), auditoryEntity.getLocation());
